packages/examples-playwright/pyproject.toml
[project]
authors = [
  { name = "A.suzuki", email = "suzu.devworks@gmail.com" },
]
dependencies = []
description = "Add your description here"
name = "examples-playwright"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = [
  "pytest-asyncio>=0.25.2",
  "pytest-cov>=6.0.0",
  "pytest-playwright>=0.6.2",
  "pytest-xdist>=3.6.1",
  "pytest>=8.3.4",
  "pytest-httpserver>=1.1.0",
  "trustme>=1.2.1",
]
